% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attcpo.R
\name{ATTCPO.OBJ}
\alias{ATTCPO.OBJ}
\title{ATTCPO.OBJ}
\usage{
ATTCPO.OBJ(y.seq, attcpo, ey1 = NULL, ey0 = NULL, attcpo.se = NULL,
  V = NULL, Y0tqteobj = NULL)
}
\arguments{
\item{y.seq}{possible values for y}

\item{attcpo}{value of attcpo}

\item{ey1}{expectation of treated potential outcomes conditional on y_{t-1}}

\item{ey0}{expectation of untreated potential outcomes conditional on y_{t-1}}

\item{attcpo.se}{standard errors for att.cpo}

\item{V}{variance matrix}

\item{Y0tqteobj}{a qte object to get the counterfactual distribution from}
}
\value{
attcpo object
}
\description{
attcpo object
}
\keyword{internal}
