% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qott.R
\name{cia.bounds}
\alias{cia.bounds}
\title{cia.bounds}
\usage{
cia.bounds(
  formla,
  xformla = ~1,
  data,
  delt.seq,
  y.seq = NULL,
  firststep = c("dr", "qr", "ll"),
  link = "logit",
  se = FALSE,
  bootiters = 100,
  cl = 1,
  alp = 0.05,
  ...
)
}
\arguments{
\item{formla}{y ~ d}

\item{xformla}{~ x1 + x2}

\item{data}{the name of the \code{data.frame}.  It should be in "long"
format rather than "wide" format.}

\item{delt.seq}{a vector of values to compute the distribution of the
treatment effect for}

\item{y.seq}{a vectof of values to compute first-step distributions over
(this is currently not used as it is computed internally)}

\item{firststep}{whether to use distribution regression ("dr"),
quantile regression ("qr"), or local linear distribution regression ("ll")
for the first step estimation of condtional distributions}

\item{link}{optional argument to pass to \code{distreg} for which link
function to use when running distribution regressions}

\item{se}{whether or not to compute standard errors (if \code{TRUE},
they are computed using the bootstrap}

\item{bootiters}{if computing standard errors using the bootstrap, how
many bootstrap iterations to use}

\item{cl}{if computing standard errors using the bootstrap, how many
cores to use in parallel computation (default is 1)}

\item{alp}{significance level for confidence intervals}

\item{...}{whatever extra arguments need to be passed to Y0tmethod}
}
\description{
Bounds on the distribution of the treatment effect and on
 the quantile of the treatment effect under a conditional independence
 assumption (cia).  It takes in a data.frame, that should indicate whether
 or not an individual is treated; separates individuals into a treated
 and untreated group; runs distribution or quantile regression to
 estimate the conditional distributions; then computes bounds on the
 DoTT or QoTT.
}
