% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attcpo.R
\name{compute.attcpo}
\alias{compute.attcpo}
\title{compute.attcpo}
\usage{
compute.attcpo(Y1t, Y0tmin1, Y0tmin2, Y0tqteobj, h = NULL, ytmin1seq = NULL,
  yseqlength = NULL, se = TRUE, iters = 100, method = "level")
}
\arguments{
\item{Y1t}{treated group outcomes in period t}

\item{Y0tmin1}{treated group outcomes in period t-1}

\item{Y0tmin2}{treated group outcomes in period t-2}

\item{Y0tqteobj}{qte object to get counterfactual quantiles from}

\item{h}{optional bandwidth}

\item{ytmin1seq}{optional sequence of values to compute attcpo(y') for}

\item{yseqlength}{optional length of sequence of values to compute
attcpo(y') for; default is 100}

\item{se}{boolean for whether or not to compute standard errors}

\item{iters}{number of bootstrap iterations to use to compute standard errors}

\item{method}{can be "level" or "rank", whether the conditional expectation
is based on the level of Y0tmin1 or its rank}
}
\description{
does the heavy lifting for computing the attcpo
}
\keyword{internal}
