% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qott.R
\name{compute.csa.bounds}
\alias{compute.csa.bounds}
\title{compute.csa.bounds}
\usage{
compute.csa.bounds(
  formla,
  t,
  tmin1,
  tmin2,
  tname,
  idname,
  data,
  delt.seq,
  y.seq,
  Y0tqteobj,
  F.y0 = NULL,
  F.y1 = NULL,
  h = NULL,
  xformla = ~1,
  firststep = c("dr", "qr", "ll"),
  cl = 1
)
}
\arguments{
\item{formla}{outcomevar ~ treatmentvar}

\item{t}{the 3rd period}

\item{tmin1}{the 2nd period}

\item{tmin2}{the 1st period}

\item{tname}{the name of the column containing periods}

\item{idname}{the name of the column containing ids}

\item{data}{a panel data frame}

\item{delt.seq}{the possible values to compute bounds on the distribution
of the treatment effect for}

\item{y.seq}{the possible values for y to take}

\item{Y0tqteobj}{a previously computed first step estimator of the
distribution of counterfactual outcomes for the treated group}

\item{F.y0}{(optional) pre-computed distribution of counterfactual untreated outcomes for the treated group}

\item{F.y1}{(optional) pre-computed distribution of treated outcomes for the treated group}

\item{h}{optional bandwidth when using local linear regression}

\item{xformla}{a formula for which covariates to use}

\item{firststep}{whether to use distribution regression ("dr"),
quantile regression ("qr"), or local linear distribution regression ("ll")
for the first step estimation of condtional distributions}

\item{cl}{(optional) number of multi-cores to use}
}
\value{
csaboundsobj
}
\description{
Compute bounds on the distribution and quantile of the
 treatment effect as given in Callaway (2017) under the copula
 stability assumption and when a first step estimator of the counterfactual
 distribution of untreated potential outcomes for the treated group is
 available.
}
