% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qott.R
\name{compute.DoTTBounds}
\alias{compute.DoTTBounds}
\title{compute.DoTTBounds}
\usage{
compute.DoTTBounds(delt.seq, y.seq, F.y1, F.y0)
}
\arguments{
\item{delt.seq}{the possible values to compute bounds on the distribution
of the treatment effect for}

\item{y.seq}{the possible values for y to take}

\item{F.y1}{(optional) pre-computed distribution of treated outcomes for the treated group}

\item{F.y0}{(optional) pre-computed distribution of counterfactual untreated outcomes for the treated group}
}
\value{
List with bounds coming from conditional distributions and Williamson-Downs bounds
}
\description{
function that does the heavy lifting for computing bounds on the DoTT (and hence
 on the QoTT) when conditional distributions of Y1t and Y0t are available
}
