% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qott.R
\name{F.Y0}
\alias{F.Y0}
\title{F.Y0}
\usage{
F.Y0(ytmin1, y.seq, Y0tmin1, Y0tmin2, Y0tqteobj, h = NULL, method = "level")
}
\arguments{
\item{ytmin1}{the value of ytmin1 to condition on}

\item{Y0tmin1}{vector of outcomes for the treated group in period t-1}

\item{Y0tmin2}{vector of outcomes for the treated group in period t-2}

\item{Y0tqteobj}{a qte object for obtaining the counterfactual distribution
of untreated potential outcomes for the treated group in period t}

\item{h}{optional bandwidth}

\item{method}{"level" or "rank" determining whether method should
be used conditional on ytmin1 or the rank of ytmin1}
}
\value{
scalar F(y|ytmin1)
}
\description{
compute F(y|ytmin1) where F is the conditional
 distribution of untreated potential outcomes for the treated group
 condtional on ytmin1;  This is computed under the copula
 stability assumption
}
\keyword{internal}
