% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attcpo.R
\name{ggattcpo}
\alias{ggattcpo}
\title{ggattcpo}
\usage{
ggattcpo(attcpoobj, ylim = NULL)
}
\arguments{
\item{attcpoobj}{an attcpo object}

\item{ylim}{optional limits of the plot}
}
\description{
plot the ATT-CPO using ggplot2
}
