% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attcpo.R, R/qott.R
\name{k}
\alias{k}
\alias{k}
\title{k}
\usage{
k(z, h = 1, type = "gaussian")

k(z, h = 1, type = "gaussian")
}
\arguments{
\item{z}{evaluate kernel at value z}

\item{h}{bandwidth}

\item{type}{either "gaussian" or "epanechnikov"}

\item{z}{evaluate kernel at}

\item{h}{bandwidth}
}
\value{
gaussian kernel

k(z/h)
}
\description{
kernel function

kernel function
}
\keyword{internal}
