% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qott.R
\name{l.inner}
\alias{l.inner}
\title{l.inner}
\usage{
l.inner(delt, y.seq, F.y1, F.y0)
}
\arguments{
\item{delt}{the value to obtain the DoTT for}

\item{y.seq}{possible values of y}

\item{F.y1}{list of distributions of treated outcomes for the treated group (length of list should be n_1)}

\item{F.y0}{list of distributions of counterfactual untreated outcomes for the treated group
(length of list should be n_1)}
}
\value{
the value of the lower bound of DoTT(
}
\description{
Obtains the lower bound on the Distribution of the
 Treatment Effect for the Treated (DoTT), DoTT(delt)
 under the Copula Stability Assumption.  These are generic in the
 sense that it is going to work without the CSA or with or without covariates.
 You just need to provide a vector of distributions -- if there are no
 covariates, all of these distributions are going to be the same
 and there is not going to be any tightening of the bounds.  If there are
 covariates or otherwise conditional distributions (e.g. coming from CSA),
 then same computational approach will work, but it will also lead to
 (potentially) tighter bounds.
}
