% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qott.R
\name{u.inner}
\alias{u.inner}
\title{u.inner}
\usage{
u.inner(delt, y.seq, F.y1, F.y0)
}
\arguments{
\item{delt}{the value to obtain the DoTT for}

\item{y.seq}{possible values of y}

\item{F.y1}{list of distributions of treated outcomes for the treated group (length of list should be n_1)}

\item{F.y0}{list of distributions of counterfactual untreated outcomes for the treated group
(length of list should be n_1)}
}
\value{
the value of the upper bound of DoTT(delt)
}
\description{
Obtains the upper bound on the Distribution of the
 Treatment Effect for the Treated (DoTT), DoTT(delt)
 under the Copula Stability Assumption
}
