% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qott.R
\name{wd.u.inner}
\alias{wd.u.inner}
\title{wd.u.inner}
\usage{
wd.u.inner(y, delt, Y1t, ddid)
}
\arguments{
\item{delt}{the value to obtain the DoTT for}

\item{Y1t}{vector of outcomes for the treated group in period t}
}
\description{
Williamson-Downs bounds inner
}
\keyword{internal}
