% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qott.R
\name{wscore}
\alias{wscore}
\title{wscore}
\usage{
wscore(bet, y, x, thisx, h)
}
\arguments{
\item{bet}{value of parameters}

\item{y}{vector of y values}

\item{x}{vector of x values}

\item{thisx}{scalar value of x for which evaluating}

\item{h}{bandwidth}
}
\value{
weighted score
}
\description{
weighted score function
}
\keyword{internal}
