% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attcpo.R
\name{E.Y1}
\alias{E.Y1}
\title{E.Y1}
\usage{
E.Y1(ytmin1val, Y1t, Y0tmin1, h = NULL, method = "level")
}
\arguments{
\item{ytmin1val}{scalar value to compute conditional expectation for}

\item{Y1t}{vector of treated potential outcomes for the treated group
in period t}

\item{Y0tmin1}{vector of untreated potential outcomes for the treated
group in period t-1}

\item{h}{optional bandwidth paramater}

\item{method}{can be "level" or "rank", whether the conditional expectation
is based on the level of Y0tmin1 or its rank}
}
\value{
the conditional expectation of y1 conditional on y0tmin1
}
\description{
a function for computing the conditional expectation
 of Y_1t given a value for Y_0tmin1
}
\examples{
data(displacements)
ytmin1 <- 10
Y1t <- subset(displacements, year==2011 & treat==1)$learn
Y0tmin1 <- subset(displacements, year==2007 & treat==1)$learn
E.Y1(ytmin1, Y1t, Y0tmin1)

}
