% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qott.R
\name{F.Y0}
\alias{F.Y0}
\title{F.Y0}
\usage{
F.Y0(
  firststep = c("dr", "qr", "ll"),
  xformla,
  data,
  yvals,
  tvals,
  h = NULL,
  retF = TRUE,
  retZ = FALSE
)
}
\arguments{
\item{firststep}{whether to use distribution regression ("dr"),
quantile regression ("qr"), or local linear distribution regression ("ll")
for the first step estimation of condtional distributions}

\item{xformla}{a formula for which covariates to use}

\item{data}{the data.frame.  It should contain colums called "Y1t" and
"Y0tmin1" which correspond to treated potential outcomes in the third
period and untreated potential outcomes in the second period}

\item{yvals}{Sequence of values to compute distributions over
(currently not used as these are determined internally)}

\item{tvals}{Sequence of values to compute conditional distribution over
(currently not used as these are determined internally)}

\item{h}{optional bandwidth when using local linear regression}

\item{retF}{whether or not to return the distribution function itself;
the default is \code{TRUE}.  If false, it returns a \code{distreg::DR} object}

\item{retZ}{whether or not to return the distribution of the transformed
random variables due to the copula stability assumption.  This is mainly
used in the numerical bootstrap procedure and, therefore, the default is
\code{FALSE}.}
}
\value{
distribution F(y|ytmin1)
}
\description{
compute F(y|ytmin1) where F is the conditional
 distribution of untreated potential outcomes for the treated group
 conditional on ytmin1.  This is computed under the copula
 stability assumption.  This function
 is typically computed internally in the \code{csabounds} package but
 is provided here for convenience.
}
