% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qott.R
\name{ggCSABounds}
\alias{ggCSABounds}
\title{ggCSABounds}
\usage{
ggCSABounds(csaboundsobj, tau = seq(0.05, 0.95, 0.05), wdbounds = FALSE,
  otherdist1 = NULL, otherdist2 = NULL)
}
\arguments{
\item{csaboundsobj}{an object returned from the csa.bounds method}

\item{tau}{vector of values between 0 and 1 to plot quantiles for}

\item{wdbounds}{boolean whether or not to also plot Williamson-Downs bounds}

\item{otherdist1}{optional ecdf of the distribution of the treatment effect
under cross sectional rank invariance}

\item{otherdist2}{optional ecdf of the distribution of the treatment effect
under panel rank invariance}
}
\description{
plot bounds on the quantile of the treatment effect
 using ggplot2
}
