% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qott.R
\name{l.inner}
\alias{l.inner}
\title{l.inner}
\usage{
l.inner(y, ytmin1, delt, ytmin1.seq, Y1t, Y0tmin1, Y0tmin2, Y0tqteobj, F.y1,
  F.y0)
}
\arguments{
\item{y}{a value to compute the inner part of Lemma 3 for}

\item{ytmin1}{the value of ytmin1 to condition on}

\item{delt}{the value to obtain the DoTT for}

\item{Y1t}{vector of outcomes for the treated group in period t}

\item{Y0tmin1}{vector of outcomes for the treated group in period t-1}

\item{Y0tmin2}{vector of outcomes for the treated group in period t-2}
}
\value{
scalar that depends on the values of y, ytmin1, and delt
}
\description{
for a particular value of y, ytmin1, delt, compute the
 max part (before taking sup) in computing the lower bound of DoTT(delt),
 see Callaway (2017), Lemma 3
}
\keyword{internal}
