% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qott.R
\name{u.inner}
\alias{u.inner}
\title{l.inner}
\usage{
u.inner(y, ytmin1, delt, ytmin1.seq, Y1t, Y0tmin1, Y0tmin2, Y0tqteobj, F.y1,
  F.y0)
}
\arguments{
\item{y}{a value to compute the inner part of Lemma 3 for}
}
\value{
scalar that depends on the values of y, ytmin1, and delt
}
\description{
for a particular value of y, ytmin1, delt, compute the
 min part (before taking sup) in computing the upper bound of DoTT(delt),
 see Callaway (2017), Lemma 3
}
\keyword{internal}
