% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qott.R
\name{wd.l}
\alias{wd.l}
\title{wd.l}
\usage{
wd.l(delt, y.seq, Y1t, ddid)
}
\arguments{
\item{delt}{the value to obtain the DoTT for}

\item{y.seq}{possible values of y}

\item{Y1t}{vector of outcomes for the treated group in period t}
}
\description{
Williamson-Downs lower bound
}
\keyword{internal}
